Bug fixes / Tying up loose ends found when deploying app to new test environment
Cleaned up Create Database code with some added missing dependencies; added some comments and conditional checks on the setup.php page to better handle missing database files. As well as comments to reduce the likelihood of future issues with code being added that doesn't properly check if the db already exists

+++ config/storesiteurlconfig.php
@@ -2,12 +2,17 @@
 If (isset($_POST['siteurlconfig'])) { 
     include_once('../shared_functions.php');
 
-    // Retrieve the value from the form submission
-    $siteurlconfig = $_POST['siteurlconfig'];
-    execquery("
-    DELETE FROM Settings Where Name = 'SiteUrlConfig'");
-    execquery("
-    INSERT INTO Settings (Name, Value) VALUES ('SiteUrlConfig', '" . $siteurlconfig . "')");
-    redirect('../setup.php');
+    If (doesDatabaseExist() && filesize($filename) > 0) {
+		// Retrieve the value from the form submission
+		$siteurlconfig = $_POST['siteurlconfig'];
+		execquery("
+		DELETE FROM Settings Where Name = 'SiteUrlConfig'");
+		execquery("
+		INSERT INTO Settings (Name, Value) VALUES ('SiteUrlConfig', '" . $siteurlconfig . "')");
+		redirect('../setup.php');
+	} else {
+		// Break! DB Doesn't exist!
+		echo "The database doesn't exist yet; please go back, and open the create DB tab and create the database.";
+		exit();
 }
 ?>
